turned editConfig.vue category HTML into wrappers (+ code formatting)
editConfig is gonna have a whole lot more stuff in it eventually, so i went with this refactor to 1. remove unnecessary line breaks 2. turn the category wrappers into a components, which makes the page structure easier to parse thanks to the different color/syntax highlighting of components versus the normal HTML