refactor/fix (docker-compose.yaml): change to relative paths
works just fine locally, but github actions is failing and logging "unable to prepare context: unable to evaluate symlinks in Dockerfile path: lstat /backend: no such file or directory"... which this refactor attempts to remedy.

googling this kind of error produced a lot of results where their path is/seems absolute, which leads me to assume that the logged path is always absolute and this error is happening because github actions' "actions/checkout@v3" is cloning to a subdir but treats this file's paths as absolute.

...i'm also guessing it might be an issue with the "/backend" path of an actual dockerfile, because "in dockerfile path: lstat /backend" is unclear (to me) on whether it's 1. docker-compose's path to the dockerfile (which is mentioned explicitly) or 2. the/a path mentioned in the/a dockerfile