test/fix (draftsAndPresets): add waiting after particular rerouting (see commit desc)
this test has been frequently failing, only on its first run, in 3 particular stages... and the error indicates that the issue occurs when failing to find an element after routing back a page.
because rerunning the test makes it pass perfectly, i speculate the issue is that on first run the page/procedure is uncached and therefore loads slowly enough that cypress proceeds before the element(s) load.
my further technical speculation is that Vue Router prematurely fires the "load" event, which (and this is according to cy docs, not speculation) cypress listens to before proceeding.

also sidenote, i've recently found out that the using cy.wait in this manner is officially an anti-pattern... but cypress' (current) official alternative is lengthier, less readable, and (seemingly) doesn't actually work in situations such as this (i.e clientside-routing delays)