refactor (UserPatchRequest): turn class into generic, narrowing newValue type
y'know how you sometimes finally figure something out and think back 'how the fuck was it so difficult to figure that out'? well, i just figured out generics...

it's extra difficult when the (typescript) docs are notoriously bad (simultaneously verbose & incomplete, so dated that errors that their code examples are throwing errors), and i don't have anyone to conceptualize this with. credit where credit is due for stackoverflow, which helped me piece it together now that i was finally motivated enough (but not also so frustrated that the tiredness trumps the motivation)