refactor(?) (about): uncomment disclaimer
this change has the chance to misinform anyone who clones and locally-deploys the project, but 1. the chances of anyone cloning this repo seems slim, and anyone who does that is likely code-literate enough (or has access to a coworker who is) to look into that 2. it's technically misinformation, but it's effect minimal and harmless 3. the home page already misinforms like that, and i'm not gonna bother with pre-hosting text for it