feat: contacts (in dashboard)
decided to go with a contacts functionality, as opposed to a friends list, due to a bunch of interpersonal preferences

this feature feels kinda incomplete without the ability to invite one's contacts during post-creation, but 1. i've yet to decide on how to design that interface (even under the current amateur standards) and 2. i'm gonna make it possible to invite people without adding them to contacts, which kinda makes post invites a distinct functionality