tests: migrate from Cypress to Playwright (doesn't work, see desc)
i have plenty of good reasons to migrate from playwright to cypress, but (and perhaps this is another good reason) it makes it very difficult/obtuse to run *completely* serialized tests... which is how i've written all of my tests thus far, and i was not aware of this situation with playwright i did all the grunt migration work.

if this was the end of the story i'd simply rewrite the tests accordingly before committing this migration, but the dashboard is currently way too delicate and needs to be tested against some wacky complications between vue proxies and state management... which, without the current serialized tests, would require a very long specialized test.

i'm soon going to rework the dashboard in a way that removes those complications (not only because of DX, but also because the resulting feature is bad UX), and then most of the tests will need to be rewritten to accommodate the dashboard changes... so, to spare myself writing a ton of temporary test code, i'm going to get the dashboard simplification done first before i migrate/fix all the tests