fix & refactor (notification): emit notif text updates, remove undefined text possibility, and replace desirability's undefined with null
with notification text wipes affecting the prop being passed in (instead of being contained within the notif component's ref), repeated notifications with the same text actually trigger the notif to reappear

removed undefined text because otherwise the most reasonable thing would seem to be to emit an change to undefined... but that would undermine typescript and likely result in some issues, as a string-only variable could be passed into the prop and then get overriden into undefined

replaced desirability's undefined with null due to an admittedly-highly-subjective conception of consistency.
also, null actually seems more appropriate than undefined because it's meant to denote a known neutral notification: a notification with undefined desirability would be informing the user either 1. that a situation is being processed, which is a known neutral state 2. that the situation resolved to an uncertain conclusion, which a user would justifiably treat as negative desirability