fix (dashboard): unbreak page rendering by removing transitions
dashboard sections started to inconsistency stop rendering after switching. after *thoroughly* trying to debug it, i failed to find any reasonable explanation for what's causing it (except for simply including recently-modified section components, even if they're empty, that makes absolutely no sense).

seems that the issue is related to how the transition wrapper, as deleting it solved the issue.
feels wrong to resolve the bug like this, but i recently realized that the dashboard sections would be better served as subroutes and when i work that refactor/feature i'll get a chance to reintroduce the transition