refactor (postContainer): move reply modal code into container component & rename picker for clarity
makes more sense for the reply interface to be layout-dependant, in some layouts (maybe even the current sole one) it could be by creating a new subnode with the inputs.

also, this would better align with the upcoming decision to make the node-editing interface layout-dependant, which in some layouts could be a modal and some layouts be an in-node replacement with nodes (or could be something else entirely).

(the rename could've been a separate commit, but i'm gonna induldge myself to not bother splitting the changes as this is currently a solo project)