refactor (createNode.vue): delete component, implementing post/reply variants where needed
for createNode.vue to also be useful in for replying, the component needed a lot of post/reply conditionals regarding the post-specific stuff (e.g config, fetch function, styling classes). a previous commit extracted the text portions (being almost all of the shared code) into a distinct editNode.vue component, which with this commit is now the main component used for replying instead of the entire createNode.vue. this allowed createNode.vue to drop all the conditionals and basically become a createPost.vue, and since it's only used the CreatePost.vue page it's been deleted and its contents subsumed (with import path adjustments) into that page compoentn