feat: dockerized dev mode (possibly excluding windows, see desc)
this *should* be working even on windows (note the usePolling addition in vite.config.ts), but for whatever reason it doesnt on my desktop. i'm committing this despite the problem so i'd be able to pull and test it on my windows laptop. first-hand confirmation that it works on linux will come in the not-too-distant future, as i'm planning to switch to linux on said laptop