docs (README.md): update test tasks
deleted the previous task because i already committed the posts feed test, and a post creation test doesn't seem worthwhile when it's covered by other tests and would duplicate specific feature tests (which i wouldn't be opposed to, considering it's a different user flow, if not for the feature density of this page making maintaining this untenable at the current stage and scale of dev)

added task for testing configs' off/absent state, after learning from the embarrassingly obvious oversight of having the "interacted" timestamp functionality on by default. gonna start implementing this type of test on each config's spec

added task for testing the "interacted" timestamp because i previously postponed it (before i even had made the TODO.md file, which is why i forgot to add this as a task) because i previously wasn't sure how to test it when it depends on the passage of time. the aforementioned "interacted" blunder got me to expedite this config by testing just the API responses and "Interacted" label (i.e ignoring the iso-converted text itself e.g "Now", "4m ago"), and any upcoming timestamp-related tests will include a "TODO" reminder for testing the timestamp component itself with a dedicated test (which is being postponed in hopes that playwright component testing leaves the experimental stage by the time i do the pre-1.0 to-do cleanup)