test (votes): comprehensive coverage upgrade
adds sorely missing tests for cumulative votes and anonymous ones, along with some relevant refactors.
the anon voting test will be expanded upon further when a proper votes list is added... pending a decision on how to implement the conversion (e.g. on server or client, via a new helper or a specialized algorithm)